requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic==2.7.1
cryptography==42.0.7
python-multipart==0.0.9
